chore : #7 - move pakage
