+++ src/main/java/org/example/setting/collection/DynamicSiteSettingCollection.java
@@ -1,9 +1,11 @@
-package org.example.setting;
+package org.example.setting.collection;
 
 import java.io.IOException;
 import java.util.List;
 
 import org.example.mapper.DataMapper;
+import org.example.setting.SettingCollection;
+import org.example.setting.SiteSetting;
 
 import lombok.Getter;
